tests, outside-connectivity: change hard-coded addresses to cmd flags
The default addresses/domains used in tests for checking connectivity
to the outside world are hard-coded.
When running the tests with no internet connection,
those tests will fail.

It should be possible to run the tests without failures,
also without internet connectivity.
It could be also that the user would like to test outside connectivity
to different addresses/domains then the default.

For that, this commit changes the hard coded addresses to be defined by the
user, and to be provided as command line flags. when addresses are not provided
the relevant tests will be skipped.

Also, the default addresses that were used till now are now passed as flags in the
functest.sh script so that the tests won't get skipped by automation.

Signed-off-by: alonsadan <asadan@redhat.com>